refactor(core): Plan B2 — extract page_fetcher/structure/llms_parser to tools/core
Strangler-fig move (translator-independent — translator imports none of these):
- tools/core/web/page_fetcher.py    <- tools/summary/page_fetcher.py
- tools/core/content/structure.py   <- tools/summary/structure.py
- tools/core/seo/llms_parser.py     <- tools/summary/llms_parser.py
Old paths are now full-namespace re-export shims (public + private, object identity
preserved) so existing imports + monkeypatch points keep working unchanged.
.importlinter: core.web/content/seo added to the core-is-a-leaf contract.

DEVIATION: url_map.py NOT moved (plan listed it). Its _main() does
`from tools.summary import config` (LLMS_TXT_URL/URL_MAP_FILE) + has a
`python -m tools.summary.url_map` entry; moving it to core/seo would make core
import a consumer tool, breaking the import-linter leaf contract. Deferred; stays
in tools/summary/. Plan C will import its pure helpers from tools.summary.url_map.

Gate: lint-imports 1 kept/0 broken; moved-module tests 52 passed; shim identity
verified (incl. LlmsIndex method); full suite 425 passed / 2 pre-existing fails.
