+++ tools/core/content/__init__.py
@@ -1,0 +1,1 @@
+"""Shared content utilities (leaf): Markdown <-> Webflow RichText HTML. See docs/ARCHITECTURE.md."""
