+++ tools/core/seo/__init__.py
@@ -1,0 +1,1 @@
+"""Shared SEO utilities (leaf): llms.txt -> LlmsIndex internal-link graph. See docs/ARCHITECTURE.md."""
