+++ tools/core/web/__init__.py
@@ -1,0 +1,1 @@
+"""Shared web utilities (leaf): live-page fetch -> PageContent. See docs/ARCHITECTURE.md."""
